refactor: rename of card component to index
