+++ src/components/Card/BigCard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Card } from './Card'
+import { Card } from './index'
 
 const meta: Meta<typeof Card.Big> = {
   title: 'common/Card/Big',
