+++ src/components/Card/NewCard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Card } from './Card'
+import { Card } from './index'
 
 const meta: Meta<typeof Card.New> = {
   title: 'common/Card/New',
