+++ src/components/Card/SmallCard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Card } from './Card'
+import { Card } from './index'
 
 const meta: Meta<typeof Card.Small> = {
   title: 'common/Card/Small',
